update tab-space and typo of ShiftPage folder
