+++ .prettierrc.json
@@ -13,7 +13,7 @@
     "quoteProps": "as-needed",
     "requirePragma": false,
     "semi": true,
-    "tabWidth": 4,
+    "tabWidth": 2,
     "useTabs": false,
     "vueIndentScriptAndStyle": false
 }
