+++ src/main.jsx
@@ -10,51 +10,51 @@ import store from './store/store';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import { BrowserRouter } from 'react-router-dom';
 import MainPage from '@pages/main/MainPage';
-import ShiftPage from '@pages/SiftPage/shiftPage';
+import ShiftPage from '@pages/ShiftPage/shiftPage';
 import StaffPage from '@pages/Staff/StaffPage';
 const router = createBrowserRouter([
-    {
-        path: '/',
-        element: <MainPage />,
-    },
-    {
-        path: '/user-info',
-        element: <MainPage />,
-    },
-    {
-        path: '/shifts',
-        element: <ShiftPage />,
-    },
-    {
-        path: '/calendar',
-        element: <MainPage />,
-    },
-    {
-        path: '/shifts',
-        element: <MainPage />,
-    },
-    {
-        path: '/collecting-sites',
-        element: <MainPage />,
-    },
-    {
-        path: '/staffs',
-        element: <StaffPage />,
-    },
-    {
-        path: '/settings',
-        element: <MainPage />,
-    },
-    {
-        path: '/logout',
-        element: <MainPage />,
-    },
+  {
+    path: '/',
+    element: <MainPage />,
+  },
+  {
+    path: '/user-info',
+    element: <MainPage />,
+  },
+  {
+    path: '/shifts',
+    element: <ShiftPage />,
+  },
+  {
+    path: '/calendar',
+    element: <MainPage />,
+  },
+  {
+    path: '/shifts',
+    element: <MainPage />,
+  },
+  {
+    path: '/collecting-sites',
+    element: <MainPage />,
+  },
+  {
+    path: '/staffs',
+    element: <StaffPage />,
+  },
+  {
+    path: '/settings',
+    element: <MainPage />,
+  },
+  {
+    path: '/logout',
+    element: <MainPage />,
+  },
 ]);
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-    <React.StrictMode>
-        <Provider store={store}>
-            <RouterProvider router={router} />
-        </Provider>
-    </React.StrictMode>,
+  <React.StrictMode>
+    <Provider store={store}>
+      <RouterProvider router={router} />
+    </Provider>
+  </React.StrictMode>,
 );
