+++ src/pages/ShiftPage/shiftPage.module.scss
@@ -1,0 +1,220 @@
+.container {
+  position: relative;
+  right: 0;
+  width: 80%;
+}
+
+.header {
+  height: 10%;
+  display: flex;
+  justify-content: center;
+}
+
+.inner {
+  height: 100%;
+  width: 90%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.new-shift-button {
+  background-color: #018241; /* Green background */
+  border: none; /* No border */
+  color: white; /* White text */
+  padding: 6px 16px; /* Padding */
+  text-align: center; /* Centered text */
+  text-decoration: none; /* No underline */
+  display: inline-block; /* Display as inline-block */
+  font-size: 16px; /* Font size */
+  margin: 40px 2px; /* Margin */
+  transition-duration: 0.4s; /* Transition time */
+  cursor: pointer; /* Cursor on hover */
+  border-radius: 50px; /* Rounded edges */
+
+  &:hover {
+    background-color: #3e8e41; /* Darker green on hover */
+    color: white; /* White text on hover */
+  }
+
+  svg {
+    margin-right: 8px;
+    color: #15f901;
+  }
+}
+
+.content {
+  display: flex;
+  height: 90%;
+  justify-content: center;
+}
+
+.inner-content {
+  height: 100%;
+  width: 90%;
+  display: block;
+}
+
+.filter-section {
+  margin-top: 0px;
+  padding: 10px 24px;
+  padding-left: 0px;
+  padding-bottom: 0px;
+  width: 100%;
+  height: 25%;
+}
+
+.filter-header {
+  width: 100%;
+  height: 16%;
+  display: flex;
+  flex-direction: row;
+  gap: 10px;
+}
+
+.filter-title {
+  height: 100%;
+  margin: 2px 0px;
+}
+
+.filter-icon {
+  size: 4px;
+  height: 100%;
+  margin: 2px 0px;
+}
+
+.filter {
+  display: flex;
+  position: relative;
+}
+
+.filter-option {
+  position: relative;
+  left: 0px;
+  height: 84%;
+  width: 48%;
+  padding: 2px 8px;
+  padding-left: 0px;
+}
+
+.filter-input {
+  position: relative;
+  left: 0;
+  flex: 1;
+  display: flex;
+  width: 100%;
+  gap: 10px;
+  margin: 15px 0px;
+  border: none;
+  box-shadow: 0 0 0 1px rgba(111, 105, 105, 0.5);
+  border-radius: 10px;
+
+  &:hover {
+    box-shadow: 0 0 0 1px #333;
+  }
+
+  &:focus-within {
+    box-shadow: 0 0 0 2px #018241;
+  }
+}
+
+.input-container {
+  height: 40px;
+  padding: 8px 0px;
+  position: relative;
+  width: 80%;
+
+  label {
+    background-color: #ffffff;
+    margin-top: 2px;
+    position: absolute;
+    top: -15px;
+    left: 20px;
+  }
+
+  input {
+    width: 100%;
+  }
+}
+
+.date-input {
+  border: none;
+  background-color: #ffff;
+  margin-left: 12px;
+  color: #000;
+
+  &:hover {
+    border: none;
+  }
+
+  &:focus {
+    outline: none;
+  }
+}
+
+.icon-container {
+  height: 40px;
+  padding: 8px 0px;
+  position: absolute;
+  right: 12px;
+}
+
+.status-input {
+  width: 100%;
+  margin: 1.5px 0;
+  border-radius: none;
+  box-shadow: none !important;
+  border: transparent;
+}
+
+.selected-status {
+  border: none !important;
+
+  border-radius: 10px !important;
+  box-shadow: none !important;
+}
+
+.filter-second-date {
+  position: relative;
+  padding: 2px 8px;
+  width: 50%;
+  text-align: right;
+}
+
+.filter-btn {
+  background-color: #018241; /* Green background */
+  border: none; /* No border */
+  color: white; /* White text */
+  padding: 6px 16px; /* Padding */
+  text-align: center; /* Centered text */
+  text-decoration: none; /* No underline */
+  display: inline-block; /* Display as inline-block */
+  font-size: 16px; /* Font size */
+  width: 128px;
+  transition-duration: 0.4s; /* Transition time */
+  cursor: pointer; /* Cursor on hover */
+  border-radius: 10px; /* Rounded edges */
+
+  &:hover {
+    background-color: #3e8e41; /* Darker green on hover */
+    color: white; /* White text on hover */
+  }
+
+  svg {
+    margin-right: 8px;
+    color: #15f901;
+  }
+}
+
+.table-section {
+  position: relative;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 60%;
+  margin-top: 12px;
+  box-shadow: 0px 24px 65px rgba(20, 20, 43, 0.16);
+  border-bottom-left-radius: 10px;
+  border-bottom-right-radius: 10px;
+  background-color: #ffffff;
+}
