+++ src/pages/ShiftPage/utils/dropdownMenu.module.scss
@@ -1,0 +1,30 @@
+.wrapper {
+  height: 5px;
+
+  .button {
+    height: 5px;
+  }
+}
+
+.toggle-btn {
+  height: 5px;
+  padding: 5px 5px;
+
+  button {
+    background-color: transparent !important;
+    border: none !important;
+    color: #018241;
+    height: 20px !important;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    &:hover {
+      color: #018241;
+    }
+
+    &::after {
+      display: none;
+    }
+  }
+}
