+++ src/pages/Staff/StaffPage.jsx
@@ -3,7 +3,7 @@ import { StaffView } from '@components/StaffView/StaffView';
 
 const StaffPage = () => {
   return (
-    <div className='container row'>
+    <div className="container row">
       <Sidebar />
       <div>
         <StaffView />
